Roll back custom binary operator typing
The problem is that our `fself` approach was broken. We would get an exact
recursive call to our current type visitor, which is not what we want. What we
actually want is to call the *outermost* function: in our case, the elaborator
mixin. I can't think of a way to accomplish that.